requirements.txt
click==7.1.2
Flask==1.1.2
future==0.18.2
gunicorn==20.0.4
h5py==2.10.0
itsdangerous==1.1.0
Jinja2==2.11.2
MarkupSafe==1.1.1
numpy==1.18.4
Pillow==7.1.2
six==1.15.0
https://download.pytorch.org/whl/cpu/torch-1.4.0%2Bcpu-cp36-cp36m-linux_x86_64.whl
https://download.pytorch.org/whl/cpu/torchvision-0.5.0%2Bcpu-cp36-cp36m-linux_x86_64.whl
Werkzeug==1.0.1
gunicorn==20.0.4
